Break the main file up into smaller required files
